+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/KeyTermSDJpaService.java
@@ -1,0 +1,16 @@
+package ru.dpopkov.knowthenics.services.springdatajpa;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+import ru.dpopkov.knowthenics.model.KeyTerm;
+import ru.dpopkov.knowthenics.repositories.KeyTermRepository;
+import ru.dpopkov.knowthenics.services.KeyTermService;
+
+@Service
+@Profile("spring-data-jpa")
+public class KeyTermSDJpaService extends AbstractSDJpaService<KeyTerm> implements KeyTermService {
+
+    public KeyTermSDJpaService(KeyTermRepository keyTermRepository) {
+        super(keyTermRepository);
+    }
+}